Moved `RandomSkillPreferences` to Skills Package
### Summary:
- Relocated `RandomSkillPreferences` from `mekhq.campaign` to `mekhq.campaign.personnel.skills` package to improve organization.

This change is purely a refactor and does not add/remove functionality or alter behavior.
