+++ MekHQ/src/mekhq/campaign/personnel/generator/AbstractPersonnelGenerator.java
@@ -36,12 +36,11 @@ import megamek.common.Compute;
 import megamek.common.enums.Gender;
 import mekhq.Utilities;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.RandomSkillPreferences;
+import mekhq.campaign.personnel.skills.RandomSkillPreferences;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.skills.SkillType;
 import mekhq.campaign.personnel.enums.PersonnelRole;
 import mekhq.campaign.personnel.enums.Phenotype;
-import mekhq.campaign.personnel.skills.SkillType;
 
 /**
  * Represents a class which can generate new {@link Person} objects for a {@link Campaign}.
