+++ MekHQ/src/mekhq/campaign/randomEvents/RoninOffer.java
@@ -57,12 +57,11 @@ import megamek.common.enums.SkillLevel;
 import megamek.common.options.IOption;
 import mekhq.Utilities;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.RandomSkillPreferences;
+import mekhq.campaign.personnel.skills.RandomSkillPreferences;
 import mekhq.campaign.finances.Money;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.skills.Skill;
 import mekhq.campaign.personnel.enums.PersonnelRole;
-import mekhq.campaign.personnel.skills.Skill;
 import mekhq.campaign.randomEvents.personalities.PersonalityController.PersonalityTraitType;
 import mekhq.campaign.randomEvents.personalities.enums.Aggression;
 import mekhq.campaign.randomEvents.personalities.enums.Ambition;
