+++ MekHQ/unittests/mekhq/campaign/universe/enums/CompanyGenerationMethodTest.java
@@ -42,7 +42,7 @@ import org.junit.jupiter.api.Test;
 import mekhq.MekHQ;
 import mekhq.campaign.Campaign;
 import mekhq.campaign.RandomOriginOptions;
-import mekhq.campaign.RandomSkillPreferences;
+import mekhq.campaign.personnel.skills.RandomSkillPreferences;
 import mekhq.campaign.universe.companyGeneration.CompanyGenerationOptions;
 import mekhq.campaign.universe.generators.companyGenerators.AtBCompanyGenerator;
 import mekhq.campaign.universe.generators.companyGenerators.WindchildCompanyGenerator;
